requirements.txt
# Core LLM
groq>=0.11.0

# Embeddings
sentence-transformers>=3.0.0

# Vector Store
chromadb>=0.5.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
python-pptx>=1.0.0
openpyxl>=3.1.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.10

# Utility
pydantic>=2.9.0
python-dotenv>=1.0.0
aiofiles>=24.0.0
numpy>=1.26.0
